common/async: `io_context_pool` constructor takes int64_t
Just because this is what Ceph's config uses and it saves a narrowing
conversion. If we want to set a max value on the thread count, we
should do it in config.

+++ src/common/async/context_pool.h
@@ -46,7 +46,7 @@ class io_context_pool {
   }
 public:
   io_context_pool() noexcept {}
-  io_context_pool(std::int16_t threadcnt) noexcept {
+  io_context_pool(std::int64_t threadcnt) noexcept {
     start(threadcnt);
   }
   ~io_context_pool() {
